meta-nuvoton: add npcm7xx-igps recipe
Fetch the XML files which describe how to assemble the bootloaders for
an NPCM7XX BMC.

Apply a patch which changes the image paths to work with our BitBake
build environment. The original files are intended to be used with
custom Python build scripts which place the intermediate images at
different paths than we use.

Change-Id: Ica2703ddf96078ff033e6febea03f1d4bb8d8126
Signed-off-by: Benjamin Fair <benjaminfair@google.com>